New model for the app and concepts
The app now has a top-level app.ttl file in /public/itme/online.

This app file contains something like this:

```
 #app ITME.hasWorkspace #2359u2345
 #2359u2345 WS.preferencesFile "./workspace/prefs.ttl"
```

the last line points to the preferences file of the workspace - there
can now be multiple workspaces in an app.

a workspace preferences file looks something like this:

```
 #workspace ITME.conceptIndex /concepts.ttl
            ITME.defaultNoteStorage "./notes"
```

New notes are created in the ITME.defaultNoteStorage container.

The concept index looks something like this:

```
 #54rcWpgERFfFqC ITME.refs #4tCWnRBwTF
                 ITME.storedAt "./notes/54rcWpgERFfFqC.ttl#concept"
 #4tCWnRBwTF ITME.storedAt "./notes/4tCWnRBwTF.ttl#concept"
```

and serves as an index of which "concepts" refer to which other
concepts. Each concept also records where it is stored, and this
storage can be changed. We call the place where a concept's body is
stored a "note"

Each note is a document that looks something like this:

```
 #concept ITME.noteBody "[{document: []}]"
```