Gate deploy (#39)
The code changes are mostly limited to one new file, /pages/api/setup-gnome.js. That maps to a vercel serveless endpoint at /api/setup-gnome. That endpoint is a POST which takes a url, the url of the gnomeThing to be deployed.

That file is then read to determine what template to deploy. A random readable Id is generated and then requested config is stored in a supabase table with url as the primary key. We use that readable id to create a github project by that name in the understory-gnomes github org. I also set up thisismytestorg for testing and preview environments, so the production git org will always be separate from the testing.

After creating the github project, we setup a new vercel project with the same name, and configure it for deployment. There are currently three steps to deploy a new project:

* Set framework to nextjs
* Set up ENV variables
* Push Empty Commit to trigger deploy

The config is saved again to supabase once the vercel and github projects have been created, then it is returned to the caller as JSON.

This PR was squashed from the following commits:

* Add basic gnomes api structure (wip)

* Plan out the basic gnome structure

* Setup github account framework.

* Add github api functions

* Add vercel/client and octokit/core

* First pass gnomesRepo creation from template

* Setup config.json in the root of the cloned template repo

* Generate and setup Vercel token

* Notes from Michiel about node auth

One idea that I had while talking with him was that we might want to
create a machine user pod as well, and then we can use the solid .acl
model to allow that user to see it.

* Updates from meetings with Michiel and Travis

- added some suggested links from Michiel about how to make
this authenticated.
- solidified intial api after conversation with Travis
- use base58 to generate a repoID from the gnomeConfigURL

* First pass at readPublicConfig using inrupt/solid-client

* Configure env variables in vercel

* Fix minor syntax issues with octokit.request

* Forgot trailing commas

* Rename api endpoint

* Use correct octokit library

* Actually send a response to the api requests

* Fix bugs in readPublicGnomeConfig

Now we can successfully read the gnome config from a users solid pod.

* Add correct error handling to findGnomesRepo

* Fix github repo creation bug

* Pass around full config instead of loose variables

* Clean up github api return values

* Add node-fetch and greg

* Successfully create Vecel Projects

* Spell it Id not ID. It's not initials.

* Add Vercel teamId to the API requests

* Add env var configuration to new vercel projects

* Add empty commit to trigger redeploy of vercel

* Spec out beginings of frotend configs for gnomes.

* Changes

* Spec out fe config after conversation with travis

* Configure the Vercel framework correctly.

* WIP attempt at doing frontend

* WIP pairing with Travis

* Working, renderable settings page with section headers and note picker

* Get basic editor working after pairing with Travis

now on to fixing that pesky save button

* Pair with Travis on basic functionality of settings page

* get new gate config on settings page looking ok

that'll do

* refactoring and tweaks towards gate deployment

* Add supabase database

* tweaks to get single page gate working

* set usesConceptIndex
* use the "public" concept prefix rather than the internal one

* Clean up and streamlining after supabase refactor

* Don't mutate config

* Fix typos

Co-authored-by: Travis Vachon <travis.vachon@gmail.com>