Fix image resizing
I'm not thrilled about this as it's not pretty and feels janky, but I could not for the life of me figure out how to use the react-dnd api to replicate this. Add stopPropagations to get react-dnd off the case, and don't allow negative widths to fix an edge case where newWidth is a negative number in the last onDrag event.