Initial notes on potential new index format
Based on existing ontologies, including SKOS:Concept. The SKOS ontology
will give us lots of useful predicates for noting things are related,
labeled, or in a particular Collection. It will also allow us to easily
specifiy notes (which can be any format, not just text documents), so we
can relate these Concepts easily to both our notes, as well as Images
and Links. This change has been prompted by the need to add Links and
Images to a garden without having an assocaited label or name yet.